move general utility functions to their own file
